refactor: move `isEmail` utility function from `frontend` to `common` so it can be used also in `mobile`
