+++ packages/common/src/utils/isEmail.ts
